requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

audiomentations == 0.12.1
hydra_core == 1.0.3
matplotlib == 3.3.2
numpy == 1.19.2
omegaconf == 2.0.2
pytorch_lightning == 1.0.4
scikit_learn == 0.23.2
torch == 1.6.0
torchaudio == 0.6.0
wandb == 0.10.8
